Add new browser support
